requirements.txt
# Core Dependencies
anthropic>=0.39.0
google-generativeai>=0.8.3
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.3.1
faiss-cpu>=1.9.0

# Web Search & Scraping
tavily-python>=0.5.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
requests>=2.32.3

# Academic Paper Processing
arxiv>=2.1.3
PyMuPDF>=1.25.1  # for PDF parsing
grobid-client-python>=0.0.7

# Machine Learning & Deep Learning
torch>=2.5.1
torchvision>=0.20.1
transformers>=4.46.3
diffusers>=0.31.0
clip-by-openai>=1.0

# Audio Processing
librosa>=0.10.2
essentia>=2.1b6.dev1110

# Image Processing
opencv-python>=4.10.0
pillow>=11.0.0

# Data Processing & Visualization
pandas>=2.2.3
numpy>=2.1.3
matplotlib>=3.9.2
seaborn>=0.13.2
scikit-learn>=1.5.2
umap-learn>=0.5.7

# Git Integration
gitpython>=3.1.43
requests>=2.32.3  # For GitHub API calls

# CLI & Utilities
rich>=13.9.4
typer>=0.15.1
tqdm>=4.67.1

# Testing (optional)
pytest>=8.3.4
pytest-asyncio>=0.24.0
